Add files via upload
Tom And Jerry Game
In this project, we created a Tom and Jerry game that is basic and straightforward! As there are minimal restrictions, anyone may play! All you have to do is jump to keep Jerry from the trap!
How to Run this source code?
To run this source code you must download both XAMPP Control panel v3.3.0 And PhpMyAdmin 5.1.3.